% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biweight.R
\name{biweight}
\alias{biweight}
\title{Biweight function}
\usage{
biweight(value, grid, radius = NULL, normalize = TRUE, var = NULL, id = NULL)
}
\arguments{
\item{value}{sf table of input values with a geometry column 'point'}

\item{grid}{sf table of grid points}

\item{radius}{if character, name of the radius column in value
of row in value else numeric constant}

\item{normalize}{if TRUE, ensure sum of pond = 1 else no correction}

\item{var}{names of 'value' variables in grid, by default all numeric
except 'id' variables}

\item{id}{excluded variables from 'var'}
}
\description{
Biweight function
}
