% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfgrid}
\alias{sfgrid}
\title{Grille sf test}
\format{
An object of class \code{sf} (inherits from \code{data.table}, \code{data.frame}) with 4 rows and 2 columns.
}
\usage{
sfgrid
}
\description{
Grille sf test
}
\keyword{datasets}
