% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfvalue}
\alias{sfvalue}
\title{Value sf test}
\format{
An object of class \code{sf} (inherits from \code{data.table}, \code{data.frame}) with 8 rows and 5 columns.
}
\usage{
sfvalue
}
\description{
Value sf test
}
\keyword{datasets}
