% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biweight.R
\name{biweight}
\alias{biweight}
\title{Biweight function}
\usage{
biweight(value, grid, radius = NULL, normalize = TRUE, var = NULL)
}
\arguments{
\item{value}{sf table of input values with a geometry column 'point'}

\item{grid}{sf table of grid points}

\item{radius}{If character, name of the radius column in value else numeric constant}

\item{normalize}{If TRUE, the defaut, ensure sum of pond = 1 else no correction}

\item{var}{Names of 'value' variables in grid, by default all numeric variables}

\item{id}{Excluded variables from 'var'}
}
\description{
Biweight function
}
