% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biweight.R
\name{biweight}
\alias{biweight}
\title{Biweight fuction}
\usage{
biweight(value, grid, id, radius, ind_normalize)
}
\arguments{
\item{value}{: sf values + points}

\item{grid}{: sf points}

\item{id}{: names of id variable in grid}

\item{radius}{: 20000 for 20 km}

\item{ind_normalize}{: if TRUE, ensure sum of pond = 1}
}
\description{
Biweight fuction
}
