% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biweight.R
\name{biweight_num}
\alias{biweight_num}
\title{Biweight raw function}
\usage{
biweight_num(input, grid, value, radius, normalize = TRUE)
}
\arguments{
\item{input}{2 cols x,y numeric matrix of value coordinates}

\item{grid}{2 cols x,y numeric matrix of grid coordinates}

\item{value}{matrix of values}

\item{radius}{numeric vector of radius (length 1 or same as x and y)}

\item{normalize}{if TRUE, ensure sum of pond = 1 else no correction}
}
\description{
Biweight raw function
}
