% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biweight.R
\name{biweight_num}
\alias{biweight_num}
\title{Biweight raw function}
\usage{
biweight_num(
  grid.x,
  grid.y,
  value.x,
  value.y,
  value.matrix,
  radius,
  normalize = TRUE
)
}
\arguments{
\item{grid.x}{= numeric vector of x grid coordinates}

\item{grid.y}{= numeric vector of y grid coordinates}

\item{value.x}{= numeric vector of x value coordinates}

\item{value.y}{= numeric vector of y value coordinates}

\item{value.matrix}{= matrix of values}

\item{radius}{: radius in same unit as x, y}

\item{normalize}{: if TRUE, ensure sum of pond = 1 else no correction}
}
\description{
Biweight raw function
}
